requirements.txt
absl-py==2.1.0
alabaster==1.0.0
astor==0.7.1
astunparse==1.6.3
babel==2.16.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.0
cycler==0.12.1
dask==2024.10.0
decorator==5.1.1
docutils==0.21.2
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
gast==0.4.0
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
idna==3.10
imageio==2.36.0
imagesize==1.4.1
importlib_metadata==8.5.0
Jinja2==3.1.4
joblib==1.4.2
keras==2.13.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
locket==1.0.0
Markdown==2.6.11
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.45.1
opencv-contrib-python
opencv-python-headless
opt_einsum==3.4.0
optree==0.13.0
osrf-pycommon==2.0.2
packaging==24.2
partd==1.4.2
pillow==11.0.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
resources==0.0.1
rich==13.9.4
rsa==4.9
scikit-image
scikit-learn
scipy
six==1.16.0
snowballstemmer==2.2.0
spectral
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
staty==1.2.4
sympy==1.13.1
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-addons==0.23.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tomli==2.0.2
toolz==1.0.0
triton==3.1.0
typeguard==2.13.3
typing_extensions==4.5.0
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.16.0
zipp==3.20.2
